% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{biodata}
\alias{biodata}
\title{Random biomarker data}
\format{
Data.frame of 100 rows and 2 columns. 
'outcome' is a factor with levels "Good" and "Poor".
'biomarker' is a numeric vector.
}
\source{
Randomly generated data
}
\usage{
biodata
}
\description{
Random biomarker data
}
\keyword{datasets}
