% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_calc_rroc.R
\name{calc_rroc_helper_nonreactive}
\alias{calc_rroc_helper_nonreactive}
\title{Helper function to calculate RROC (Receiver Operating Characteristic) curves}
\usage{
calc_rroc_helper_nonreactive(
  rroc_fun,
  dvs,
  ivs,
  positive_label,
  rroc_results,
  data00,
  n_permutations,
  recalculate_rroc
)
}
\arguments{
\item{rroc_fun}{A function to calculate RROC.}

\item{dvs}{A character vector of dependent variable names.}

\item{ivs}{A character vector of independent variable names.}

\item{positive_label}{A character string indicating the label for the positive class.}

\item{rroc_results}{A list containing the current RROC results.}

\item{data00}{A data frame used for calculations.}

\item{n_permutations}{A numeric value indicating the number of permutations to perform.}

\item{recalculate_rroc}{A logical value indicating whether to recalculate RROC curves.}
}
\value{
A list with updated RROC calculations.
}
\description{
This function calculates RROC curves for given dependent and independent variables.
It supports recalculating RROC curves if needed.
}
\details{
The function first checks if `rroc_results` is NULL. If it is, it calculates the RROC curves
for all combinations of dependent and independent variables. If `rroc_results` is not NULL,
it either recalculates the RROC curves for all combinations or only for new combinations
that have not been calculated before, based on the value of `recalculate_rroc`.
}
\examples{
\dontrun{
calc_rroc_helper_nonreactive(
    rroc_fun = my_rroc_function,
    dvs = c("dependent_var1", "dependent_var2"),
    ivs = c("independent_var1", "independent_var2"),
    positive_label = "positive",
    rroc_results = list(),
    data00 = data.frame(),
    n_permutations = 100,
    recalculate_rroc = TRUE
)
}
}
