% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{download_restriction_performance}
\alias{download_restriction_performance}
\title{Download restriction performance}
\usage{
download_restriction_performance(restriction_perf, fname)
}
\arguments{
\item{restriction_perf}{Reactive value for restriction performance}

\item{fname}{Filename for the downloaded file}
}
\value{
A downloadable file
}
\description{
Download restriction performance
}
