% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{glehr2023_cd4_cd8_relative}
\alias{glehr2023_cd4_cd8_relative}
\title{Cell proportions and irAEs of 177 samples from Glehr et al. (2023)
 
See the original publication for more details. The data comes from the R package dataMelanoma (https://github.com/ggrlab/dataMelanoma).}
\format{
Tibble of 177 rows and 89 columns. 
Contains sample ids, hepatitis and colitis status, presence in the validation set of an earlier 
publication and the proportion or count of 84 cell types measured by flow cytometry.
}
\source{
Randomly generated data
}
\usage{
glehr2023_cd4_cd8_relative
}
\description{
Cell proportions and irAEs of 177 samples from Glehr et al. (2023)
 
See the original publication for more details. The data comes from the R package dataMelanoma (https://github.com/ggrlab/dataMelanoma).
}
\keyword{datasets}
