% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_report_rroc.R
\name{has_been_calculated}
\alias{has_been_calculated}
\title{Check if RROC has been calculated}
\usage{
has_been_calculated(dv, iv, rroc_result)
}
\arguments{
\item{dv}{Dependent variable}

\item{iv}{Independent variable}

\item{rroc_result}{Reactive value for RROC result}
}
\value{
TRUE if RROC has been calculated, FALSE otherwise
}
\description{
Check if RROC has been calculated
}
