% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{observe_clipboard_data}
\alias{observe_clipboard_data}
\title{Observe clipboard data}
\usage{
observe_clipboard_data(input, current_data)
}
\arguments{
\item{input}{Shiny input object}

\item{current_data}{Reactive value for current data}
}
\description{
Observe clipboard data
}
