% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{observe_data_upload}
\alias{observe_data_upload}
\title{Observe data upload}
\usage{
observe_data_upload(input, current_data, rroc_result)
}
\arguments{
\item{input}{Shiny input object}

\item{current_data}{Reactive value for current data}

\item{rroc_result}{Reactive value for RROC result}
}
\description{
Observe data upload
}
