% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{observe_dv_iv_selection}
\alias{observe_dv_iv_selection}
\title{Observe DV and IV selection}
\usage{
observe_dv_iv_selection(input, session, dv_selector, iv_selector, all_cols)
}
\arguments{
\item{input}{Shiny input object}

\item{session}{Shiny session object}

\item{dv_selector}{Reactive value for DV selector}

\item{iv_selector}{Reactive value for IV selector}

\item{all_cols}{Reactive value for all columns}
}
\description{
Observe DV and IV selection
}
