% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{observe_rroc_calculation}
\alias{observe_rroc_calculation}
\title{Observe RROC calculation}
\usage{
observe_rroc_calculation(
  input,
  current_data,
  rroc_result,
  possible_positive_labels
)
}
\arguments{
\item{input}{Shiny input object}

\item{current_data}{Reactive value for current data}

\item{rroc_result}{Reactive value for RROC result}

\item{possible_positive_labels}{Reactive value for possible positive labels}
}
\description{
Observe RROC calculation
}
