% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_report_rroc.R
\name{observe_rroc_plot_update}
\alias{observe_rroc_plot_update}
\title{Observe RROC plot update}
\usage{
observe_rroc_plot_update(
  rroc_result,
  data00,
  redo_plot,
  listen_iv_dv_first,
  positive_label
)
}
\arguments{
\item{rroc_result}{Reactive value for RROC result}

\item{redo_plot}{Reactive value for redo plot}

\item{listen_iv_dv_first}{Reactive value for listening to first DV and IV}

\item{input}{Shiny input object}

\item{current_data}{Reactive value for current data}
}
\description{
Observe RROC plot update
}
