% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_preview.R
\name{previewDataUI}
\alias{previewDataUI}
\title{Preview data}
\usage{
previewDataUI(id)
}
\arguments{
\item{id}{A character string representing the module's ID.}
}
\value{
A Shiny module for previewing data.
}
\description{
This module provides a preview of the first 6 columns of a dataset.
}
