% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{render_data_preview}
\alias{render_data_preview}
\title{Render data preview}
\usage{
render_data_preview(input, current_data_table)
}
\arguments{
\item{input}{Shiny input object}

\item{current_data_table}{Reactive value for current data table}
}
\value{
A data table preview
}
\description{
Render data preview
}
