% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_newdata.R
\name{render_data_ui}
\alias{render_data_ui}
\title{Render data UI}
\usage{
render_data_ui(id, possible_data_types = c("csv", "clipboard", "glehr2023"))
}
\value{
A UI element for data input
}
\description{
Render data UI
}
