% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{render_dvs_ui}
\alias{render_dvs_ui}
\title{Render dependent variables UI}
\usage{
render_dvs_ui(all_cols)
}
\arguments{
\item{all_cols}{Reactive value for all columns}
}
\value{
A UI element for selecting dependent variables
}
\description{
Render dependent variables UI
}
