% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{render_ivs_ui}
\alias{render_ivs_ui}
\title{Render independent variables UI}
\usage{
render_ivs_ui(all_cols, input)
}
\arguments{
\item{all_cols}{Reactive value for all columns}

\item{input}{Shiny input object}
}
\value{
A UI element for selecting independent variables
}
\description{
Render independent variables UI
}
