% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{render_positive_labels_ui}
\alias{render_positive_labels_ui}
\title{Render positive labels UI}
\usage{
render_positive_labels_ui(possible_positive_labels)
}
\arguments{
\item{possible_positive_labels}{Reactive value for possible positive labels}
}
\value{
A UI element for selecting positive labels
}
\description{
Render positive labels UI
}
