% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{render_rroc_plot}
\alias{render_rroc_plot}
\title{Render RROC plot}
\usage{
render_rroc_plot(redo_plot, rroc_result, current_data, input)
}
\arguments{
\item{redo_plot}{Reactive value for redo plot}

\item{rroc_result}{Reactive value for RROC result}

\item{current_data}{Reactive value for current data}

\item{input}{Shiny input object}
}
\value{
A plot of RROC
}
\description{
Render RROC plot
}
