% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_report_rroc.R
\name{report_rroc_ui}
\alias{report_rroc_ui}
\title{Preview data}
\usage{
report_rroc_ui(id)
}
\arguments{
\item{id}{A character string representing the module's ID.}
}
\value{
A Shiny module for previewing data.
}
\description{
This module provides a preview of the first 6 columns of a dataset.
}
