% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{uploadfile_fun}
\alias{uploadfile_fun}
\title{Generate file input UI based on selected data type}
\usage{
uploadfile_fun(selected_data_type)
}
\arguments{
\item{selected_data_type}{The type of data selected by the user}
}
\value{
A file input UI element
}
\description{
Generate file input UI based on selected data type
}
