% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{observe_rroc_plot_update}
\alias{observe_rroc_plot_update}
\title{Observe RROC plot update}
\usage{
observe_rroc_plot_update(
  input,
  rroc_result,
  current_data,
  redo_plot,
  listen_iv_dv_first
)
}
\arguments{
\item{input}{Shiny input object}

\item{rroc_result}{Reactive value for RROC result}

\item{current_data}{Reactive value for current data}

\item{redo_plot}{Reactive value for redo plot}

\item{listen_iv_dv_first}{Reactive value for listening to first DV and IV}
}
\description{
Observe RROC plot update
}
